refactor tests
* reorganize to mirror library usage
* add a bit more coverage
